# docs/swagger.yaml
basePath: /
definitions:
  subs.CreateSubscriptionRequest:
    properties:
      end_date:
        example: 12-2024
        type: string
      price:
        example: 1000
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 01-2024
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  subs.ErrorResponse:
    properties:
      message:
        example: error description
        type: string
    type: object
  subs.SubscriptionResponse:
    properties:
      end_date:
        example: 12-2024
        type: string
      id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
      price:
        example: 1000
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 01-2024
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  subs.SummaryRequest:
    properties:
      end_date:
        example: 12-2024
        type: string
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 01-2024
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  subs.SummaryResponse:
    properties:
      summary:
        example: 12000
        type: integer
    type: object
  subs.UpdateSubscriptionRequest:
    properties:
      end_date:
        example: 12-2024
        type: string
      price:
        example: 1500
        type: integer
      service_name:
        example: Netflix Premium
        type: string
      start_date:
        example: 01-2024
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Subscriptions API
  version: "1.0"
paths:
  /subs:
    post:
      consumes:
      - application/json
      description: Creates a new subscription record for a user.
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/subs.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/subs.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /subs/{id}:
    delete:
      description: Deletes a subscription by its ID.
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subscription successfully deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      description: Returns subscription details by its ID.
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subs.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
      summary: Get subscription
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Updates an existing subscription by its ID.
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/subs.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subs.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
  /subs/list/{id}:
    get:
      description: Returns all subscriptions for a specific user.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/subs.SubscriptionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
      summary: List user subscriptions
      tags:
      - subscriptions
  /subs/summary:
    post:
      consumes:
      - application/json
      description: |-
        Calculates the total amount spent on subscriptions within a date range.
        Both start_date and end_date are required; user_id and service_name are optional filters.
      parameters:
      - description: Summary request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subs.SummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subs.SummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subs.ErrorResponse'
      summary: Calculate total payments
      tags:
      - subscriptions
swagger: "2.0"